polish up setting menu (no functionality atm)
