+++ Assets/Resource/arrow_down.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b71b98b46c87a04cb9091fb299c0a34
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
