+++ Assets/Resource/arrow_left.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 264a9534802518f4686fc4999ef7d28a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
