+++ Assets/Resource/arrow_right.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 490f2472874073b42b2524ac61f4ceae
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
