+++ Assets/Resource/arrow_up.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0794fdbed405c784aaa3ae7ca8c65d3b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
