+++ Assets/Resource/setting_background.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 275e378c4847b71428ec62f1116e5545
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
