+++ Assets/Resource/volume_slider_background.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 09b05924cae75e142931bad83f9e3d92
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
