+++ Assets/Resource/volume_slider_fill_background.ai.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 785c691c6624dd04dbee312de2e7746f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
