+++ Assets/Scripts/AudioSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04d274d305be3a74686ea2d4806ff54e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
