--- Assets/StageSelectMenu.cs
@@ -1,26 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class StageSelectMenu : MonoBehaviour
-{
-    public void Movement101Button ()
-    {
-        // TODO load Movement101 level select
-    }
-    
-    public void Health101Button ()
-    {
-        // TODO load Health101 level select
-    }
-    
-    public void Damage101Button ()
-    {
-        // TODO load Damage101 level select
-    }
-
-    public void BackButton ()
-    {
-    }
-}
